// apps/apis/api-admin/src/medicina-laboral/declaracion-jurada/dto/declaracion-jurada.response.ts
export class DeclaracionJuradaResponse {
  Id: number;
  Id_empresa?: number;
  Id_Paciente?: number;
  Id_Practica?: number;
  Fecha?: Date;
  Status?: string;
  Genero?: string;
  Fuma?: number;
  Cantidad_Cigarros_diarios?: number;
  Consume_Alcohol?: number;
  Frecuencia?: string;
  Practica_Deportes?: number;
  Cuales_Deportes?: string;
  Mano_Habil?: string;
  Ultima_Tarea_Trabajo_Puesto?: string;
  Recibio_Indeminizacion_o_tiene_pendiente?: number;
  Recibio_Indeminizacion_o_tiene_pendiente_Aclaraciones?: string;
  Mareos_Desmayos?: number;
  Mareos_Desmayos_Aclaraciones?: string;
  Convulsiones?: number;
  Convulsiones_Aclaraciones?: string;
  Epilepsia?: number;
  Epilepsia_Aclaraciones?: string;
  Depresion?: number;
  Depresion_Aclaraciones?: string;
  Enfermedad_Neurologica?: number;
  Enfermedad_Neurologica_Aclaraciones?: string;
  Diabetes?: number;
  Diabetes_Aclaraciones?: string;
  Dolor_Frecuente_de_Cabeza?: number;
  Dolor_Frecuente_de_Cabeza_Aclaraciones?: string;
  Toma_Medicacion?: number;
  Toma_Medicacion_Aclaraciones?: string;
  Hipertension_Arterial?: number;
  Hipertension_Arterial_Aclaraciones?: string;
  Cardiopatias?: number;
  Cardiopatias_Aclaraciones?: string;
  Varices?: number;
  Varices_Aclaraciones?: string;
  Chagas?: number;
  Chagas_Aclaraciones?: string;
  Dentadura_Sana?: number;
  Dentadura_Sana_Aclaraciones?: string;
  Alergias?: number;
  Alergias_Aclaraciones?: string;
  Tumor_Cancer?: number;
  Tumor_Cancer_Aclaraciones?: string;
  Hepatitis?: number;
  Hepatitis_Aclaraciones?: string;
  Disminucion_Auditiva?: number;
  Disminucion_Auditiva_Aclaraciones?: string;
  Tratamiento_Tiroides?: number;
  Tratamiento_Tiroides_Aclaraciones?: string;
  Enfermedades_de_la_Piel?: number;
  Enfermedades_de_la_Piel_Aclaraciones?: string;
  Analisis_HIV?: number;
  Analisis_HIV_Aclaraciones?: string;
  Asma_Tos_Cronica?: number;
  Asma_Tos_Cronica_Aclaraciones?: string;
  Neumonia?: number;
  Neumonia_Aclaraciones?: string;
  Sangre_al_Escupir?: number;
  Sangre_al_Escupir_Aclaraciones?: string;
  Celiaquia?: number;
  Celiaquia_Aclaraciones?: string;
  Hernias?: number;
  Hernias_Aclaraciones?: string;
  Tratamiento_Psiquiatrico?: number;
  Tratamiento_Psiquiatrico_Aclaraciones?: string;
  Fracturas?: number;
  Fracturas_Aclaraciones?: string;
  Lumbalgia?: number;
  Lumbalgia_Aclaraciones?: string;
  Problema_Renal_Urinario?: number;
  Problema_Renal_Urinario_Aclaraciones?: string;
  Amputaciones?: number;
  Amputaciones_Aclaraciones?: string;
  Limitaciones_Funcionales?: number;
  Limitaciones_Funcionales_Aclaraciones?: string;
  Traumatismos_Lesiones?: number;
  Traumatismos_Lesiones_Aclaraciones?: string;
  Enfermedades_Oculares?: number;
  Enfermedades_Oculares_Aclaraciones?: string;
  Uso_de_Lentes?: number;
  Uso_de_Lentes_Aclaraciones?: string;
  Ulcera_Gastrica_Nauseas_Vomitos?: number;
  Ulcera_Gastrica_Nauseas_Vomitos_Aclaraciones?: string;
  Cirugias?: number;
  Cirugias_Aclaraciones?: string;
  Internaciones?: number;
  Internaciones_Aclaraciones?: string;
  Vacunas_COVID?: number;
  Vacunas_COVID_Aclaraciones?: string;
  Tuvo_Algun_Embarazo?: number;
  Cuantos_Embarazos?: string;
  Fecha_Ultima_Menstruacion?: Date;
  Abortos_Espontaneos?: number;
  Abortos_Espontaneos_Cuantos?: number;
  Tratamientos_Hormonales?: number;
  Tratamientos_Hormonales_Aclaraciones?: string;
  Consume_Marihuana_Aclaraciones?: string;
  Consume_Cocaina?: number;
  Consume_Cocaina_Aclaraciones?: string;
  Consume_Marihuana?: number;
}